refactor: rename layout to applayout and remove extra applayout
