+++ src/components/AppComponents/AppLayout/AppLayout.test.tsx
@@ -1,6 +1,6 @@
 import { act, screen } from "@testing-library/react";
 
-import Layout from "./Layout";
+import AppLayout from "./AppLayout";
 
 import { DEFAULT_SEO_PROPS } from "@/browser-lib/seo/Seo";
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
@@ -10,21 +10,27 @@ const render = renderWithProviders();
 
 describe("Layout", () => {
   it("renders a header", () => {
-    render(<Layout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />);
+    render(
+      <AppLayout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />,
+    );
 
     const header = screen.getByRole("banner");
 
     expect(header).toBeInTheDocument();
   });
   it("renders a main", () => {
-    render(<Layout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />);
+    render(
+      <AppLayout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />,
+    );
 
     const main = screen.getByRole("main");
 
     expect(main).toBeInTheDocument();
   });
   it("renders a hidden skip to content button until focused", () => {
-    render(<Layout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />);
+    render(
+      <AppLayout seoProps={DEFAULT_SEO_PROPS} topNavProps={topNavFixture} />,
+    );
     const skipButtonLink = screen.getAllByRole("link", {
       name: "Skip to content",
     })[0];
