+++ src/components/AppComponents/AppLayout/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./AppLayout";
+export type { AppLayoutProps as LayoutProps } from "./AppLayout";
