--- src/components/AppComponents/Layout/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Layout";
-export type { LayoutProps } from "./Layout";
