+++ src/components/GenericPagesViews/PostListing.view.tsx
@@ -22,7 +22,7 @@ import {
   getBlogWebinarListBreadcrumbs,
 } from "@/components/SharedComponents/Breadcrumbs/getBreadcrumbs";
 import SummaryCard from "@/components/SharedComponents/Card/SummaryCard";
-import Layout from "@/components/AppComponents/Layout";
+import Layout from "@/components/AppComponents/AppLayout";
 import MobileFilters from "@/components/SharedComponents/MobileFilters";
 import {
   PostListingPageProps,
