--- src/components/SharedComponents/AppLayout/AppLayout.tsx
@@ -1,12 +1,0 @@
-import { FC } from "react";
-
-import Layout, { LayoutProps } from "@/components/AppComponents/Layout";
-
-export type AppLayoutProps = LayoutProps;
-const AppLayout: FC<AppLayoutProps> = (props) => {
-  const { children, ...layoutProps } = props;
-
-  return <Layout {...layoutProps}>{children}</Layout>;
-};
-
-export default AppLayout;
