--- src/components/SharedComponents/AppLayout/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./AppLayout";
-export type { AppLayoutProps } from "./AppLayout";
