+++ src/pages/about-us/oaks-curricula.tsx
@@ -8,7 +8,7 @@ import {
 } from "@oaknational/oak-components";
 import styled from "styled-components";
 
-import Layout from "@/components/AppComponents/Layout";
+import Layout from "@/components/AppComponents/AppLayout";
 import { AboutUsLayout } from "@/components/GenericPagesComponents/AboutUsLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import {
