+++ src/pages/lesson-planning.tsx
@@ -11,7 +11,7 @@ import {
   OakLink,
 } from "@oaknational/oak-components";
 
-import Layout from "@/components/AppComponents/Layout";
+import Layout from "@/components/AppComponents/AppLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import getPageProps from "@/node-lib/getPageProps";
 import Breadcrumbs from "@/components/SharedComponents/Breadcrumbs";
